Feat: portfolio pie chart, exact decimal amounts, free month navigation
- Dashboard Portfolio now renders a sorted pie chart with legend instead
  of a plain chip list, using a curated 20-color palette for clear
  per-category differentiation.
- Reports: remove the AI Spending Insights card and its now-orphaned
  backend endpoint/service (chat and category-suggestion AI features
  are unaffected).
- Switch all monetary fields (transactions, categories, goals,
  investments) from Float to Numeric(12,2)/Decimal end-to-end, with a
  migration, to stop floating-point drift in stored and aggregated
  amounts. currencyFormat pipe now defaults to showing exact cents
  everywhere instead of rounding to whole numbers.
- Dashboard and Transactions month navigation no longer locks users to
  the current calendar month, so next month's expenses (e.g. salary
  arriving a few days early) can be entered before it actually starts.

+++ backend/app/models/investment.py
@@ -1,5 +1,6 @@
 from datetime import datetime
-from sqlalchemy import String, Float, DateTime, ForeignKey
+from decimal import Decimal
+from sqlalchemy import String, Numeric, DateTime, ForeignKey
 from sqlalchemy.orm import Mapped, mapped_column
 from app.models.base import Base
 
@@ -10,6 +11,6 @@ class Investment(Base):
     id: Mapped[int] = mapped_column(primary_key=True, index=True)
     name: Mapped[str] = mapped_column(String(100), nullable=False)
     type: Mapped[str] = mapped_column(String(20), nullable=False)  # etf, crypto, cash, stocks, real_estate
-    value: Mapped[float] = mapped_column(Float, nullable=False)
+    value: Mapped[Decimal] = mapped_column(Numeric(12, 2), nullable=False)
     updated_at: Mapped[datetime] = mapped_column(DateTime, default=datetime.utcnow, onupdate=datetime.utcnow)
     user_id: Mapped[int] = mapped_column(ForeignKey("users.id"), nullable=False, index=True)
