+++ backend/app/schemas/transaction.py
@@ -1,4 +1,5 @@
 import datetime
+from decimal import Decimal
 from typing import Optional
 from pydantic import BaseModel
 from app.schemas.category import CategoryRead
@@ -6,7 +7,7 @@ from app.schemas.category import CategoryRead
 
 class TransactionBase(BaseModel):
     date: datetime.date
-    amount: float
+    amount: Decimal
     description: Optional[str] = None
     type: str  # income | expense
     category_id: Optional[int] = None
@@ -18,7 +19,7 @@ class TransactionCreate(TransactionBase):
 
 class TransactionUpdate(BaseModel):
     date: datetime.date | None = None
-    amount: float | None = None
+    amount: Decimal | None = None
     description: str | None = None
     type: str | None = None
     category_id: int | None = None
